refactor: split adjective from word metadata
