+++ harper-core/src/dict_word_metadata/mod.rs
@@ -1,3 +1,4 @@
+pub mod adjective;
 pub mod dialect;
 pub mod nominal;
 pub mod orthography;
@@ -8,12 +9,12 @@ pub mod verb;
 pub mod tests;
 
 use harper_brill::UPOS;
-use is_macro::Is;
 use paste::paste;
 use serde::{Deserialize, Serialize};
 
 use crate::{
     dict_word_metadata::{
+        adjective::{AdjectiveData, Degree},
         dialect::DialectFlags,
         nominal::{NounData, Person, PronounData},
         orthography::OrthFlags,
@@ -571,33 +572,6 @@ impl DeterminerData {
             is_demonstrative: self.is_demonstrative.or(other.is_demonstrative),
             is_possessive: self.is_possessive.or(other.is_possessive),
             is_quantifier: self.is_quantifier.or(other.is_quantifier),
-        }
-    }
-}
-
-/// Degree is a property of adjectives: positive is not inflected
-/// Comparative is inflected with -er or comes after the word "more"
-/// Superlative is inflected with -est or comes after the word "most"
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Serialize, Deserialize, PartialOrd, Is, Hash)]
-pub enum Degree {
-    Positive,
-    Comparative,
-    Superlative,
-}
-
-/// Some adjectives are not comparable so don't have -er or -est forms and can't be used with "more" or "most".
-/// Some adjectives can only be used "attributively" (before a noun); some only predicatively (after "is" etc.).
-/// In old grammars words like the articles and determiners are classified as adjectives but behave differently.
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, PartialOrd, Eq, Hash, Default)]
-pub struct AdjectiveData {
-    pub degree: Option<Degree>,
-}
-
-impl AdjectiveData {
-    /// Produce a copy of `self` with the known properties of `other` set.
-    pub fn or(&self, other: &Self) -> Self {
-        Self {
-            degree: self.degree.or(other.degree),
         }
     }
 }
